proper fetching endpoint and updating of types
